Feature/project methods (#71)
* Methods updated (returns Promises)

* Get-projects added

* Disable eslint

* Fix app/backend code style

* Remove redundant methods

* Add missed changes after merge
