+++ app/backend/src/index.ts
@@ -1,7 +1,7 @@
 import fastify from 'fastify';
 import fastifyStatic from 'fastify-static';
 import path from 'path';
-import { createTransportServer } from './ui/ctproto';
+import { createTransportServer } from './ui/transport-server';
 import { Config } from './config/config';
 
 const server = fastify();
