+++ app/backend/src/ui/methods/projects/get-projects.ts
@@ -1,0 +1,11 @@
+import Project from '../../../../../types/entities/project';
+import ProjectSchema from '../../../database/models/project';
+
+/**
+ * Returns list of stored projects
+ */
+export async function getProjects(): Promise<Project[]> {
+  return ProjectSchema
+    .find()
+    .exec();
+}
