--- app/backend/src/ui/methods/projects/update-title.ts
@@ -1,11 +1,0 @@
-import Project from '../../../database/models/project';
-import { Query } from 'mongoose';
-
-/**
- * @param id
- * @param newTitle
- */
-export function updateTitle(id: string, newTitle: string): Query<any, any> {
-  return Project.findByIdAndUpdate(id,
-    { title: newTitle });
-}
